+++ R/as_bm_pixmap.R
@@ -44,12 +44,19 @@ as_bm_pixmap <- function(x, ...) {
 #' @rdname as_bm_pixmap
 #' @export
 as_bm_pixmap.default <- function(x, ...) {
-    as_bm_pixmap.raster(grDevices::as.raster(x, ...))
+    # "bitmap" "rgba" array from `pdftools::pdf_render_page()`
+    if (is.array(x) && is.raw(x))
+        as_bm_pixmap.array(x)
+    else
+        as_bm_pixmap.raster(grDevices::as.raster(x, ...))
 }
 
 #' @rdname as_bm_pixmap
 #' @export
 as_bm_pixmap.array <- function(x, ...) {
+    if (is.raw(x)) {
+        x <- aperm(structure(as.numeric(x) / 255, dim = dim(x)))
+    }
     as_bm_pixmap.raster(grDevices::as.raster(x, ...))
 }